Put selectors in own file
